feat: Python interface application for generation of context.jsonld
