+++ mapping_suite_sdk/tools/entrypoints/cli/convert.py
@@ -1,4 +1,5 @@
 import logging
+import shutil
 from enum import Enum
 from pathlib import Path
 
@@ -15,7 +16,6 @@ from mapping_suite_sdk.mapping_package_v3.adapters.package_serialiser_lightweigh
 from mapping_suite_sdk.mapping_package_v3.services.load_mapping_package_v3 import load_mapping_package_v3_from_folder
 from mapping_suite_sdk.tools.entrypoints.cli import typer_verbose_callback
 from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import convert_mapping_package_v2_to_v3
-from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
 from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import is_mapping_package_already_converted
 from mapping_suite_sdk.tools.services.convert_mapping_package_v3_to_v3_lightweight import convert_mapping_package_v3_to_v3_lightweight
 
@@ -127,21 +127,86 @@ def _remove_conceptual_mapping_file(mapping_package_folder_path: Path) -> None:
             message=f"Removed conceptual_mappings.xlsx file from {conceptual_mapping_path}"))
 
 
+def _find_project_root(start_path: Path) -> Path:
+    """
+    Find the project root by looking for common markers (pyproject.toml, etc.).
+    
+    Args:
+        start_path: Starting directory to search from
+        
+    Returns:
+        Path to project root
+    """
+    current = start_path.resolve()
+    
+    # Look for project markers
+    markers = ["pyproject.toml", "poetry.lock", ".git"]
+    
+    while current != current.parent:
+        for marker in markers:
+            if (current / marker).exists():
+                return current
+        current = current.parent
+    
+    # If we reach here, we couldn't find the root
+    # Fall back to assuming current working directory is project root
+    return Path.cwd()
+
+
+def _copy_context_jsonld_to_package(mapping_package_folder_path: Path, converted_package) -> None:
+    """
+    Copy context.jsonld from schema directory to the package.
+    
+    The context.jsonld file is generated in the schema directory (resources/schema/mapping_package_v3/models/)
+    when models are generated. This function copies it to the same directory as metadata.jsonld in the package.
+    
+    Args:
+        mapping_package_folder_path: Path to the mapping package folder
+        converted_package: The converted package (V3 or V3L) containing metadata with path information
+    """
+    # Find project root
+    project_root = _find_project_root(Path.cwd())
+    
+    # Schema context file location
+    schema_context_path = project_root / "resources" / "schema" / "mapping_package_v3" / "models" / "context.jsonld"
+    
+    # If context.jsonld doesn't exist in schema directory, skip (it should be generated by Makefile)
+    if not schema_context_path.exists():
+        logger.warning(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
+            package_source=mapping_package_folder_path,
+            message=f"context.jsonld not found in schema directory at {schema_context_path}. "
+                   f"Run 'make generate-models' to generate it."))
+        return
+    
+    # Determine metadata directory (where metadata.jsonld is located)
+    metadata_path = mapping_package_folder_path / converted_package.metadata.path
+    metadata_directory = metadata_path.parent
+    
+    # Destination for context.jsonld (same directory as metadata.jsonld)
+    package_context_path = metadata_directory / "context.jsonld"
+    
+    # Copy the file
+    shutil.copy2(schema_context_path, package_context_path)
+    logger.debug(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
+        package_source=mapping_package_folder_path,
+        message=f"Copied context.jsonld from {schema_context_path} to {package_context_path}"))
+
+
 def _serialise_mapping_package(to_version: str, mapping_package_folder_path: Path, converted_package):
     """Dynamically serialize a mapping package based on version."""
     if to_version == Version.V3:
         serialiser = MappingPackageV3Serialiser()
         serialiser.serialise(mapping_package_folder_path, converted_package)
-        # Generate context.jsonld for V3 packages
-        _generate_context_jsonld_for_v3(mapping_package_folder_path, converted_package)
+        # Copy context.jsonld from schema directory to package
+        _copy_context_jsonld_to_package(mapping_package_folder_path, converted_package)
         # Remove old metadata.json if it exists (V2 uses metadata.json, V3 uses metadata.jsonld)
         # Do this after serialization to ensure it's not recreated
         _remove_old_metadata_json(mapping_package_folder_path)
     elif to_version == Version.V3L:
         serialiser = MappingPackageV3LightweightSerialiser()
         serialiser.serialise(mapping_package_folder_path, converted_package)
-        # Generate context.jsonld for V3 lightweight packages
-        _generate_context_jsonld_for_v3(mapping_package_folder_path, converted_package)
+        # Copy context.jsonld from schema directory to package
+        _copy_context_jsonld_to_package(mapping_package_folder_path, converted_package)
         # Remove old metadata.json if it exists (V2 uses metadata.json, V3 uses metadata.jsonld)
         # Do this after serialization to ensure it's not recreated
         _remove_old_metadata_json(mapping_package_folder_path)
@@ -150,38 +215,6 @@ def _serialise_mapping_package(to_version: str, mapping_package_folder_path: Pat
         _remove_conceptual_mapping_file(mapping_package_folder_path)
     else:
         raise typer.BadParameter(f"Unsupported target version: {to_version}")
-
-
-def _generate_context_jsonld_for_v3(mapping_package_folder_path: Path, converted_package):
-    """
-    Generate context.jsonld file for V3 mapping packages.
-    
-    The context.jsonld file is placed in the same directory as metadata.jsonld.
-    This function is called automatically during conversion for both single packages
-    and bulk folder conversions.
-    """
-    # Determine metadata directory (where metadata.jsonld is located)
-    metadata_path = mapping_package_folder_path / converted_package.metadata.path
-    metadata_directory = metadata_path.parent
-    
-    # Schema path relative to project root
-    schema_path = Path("resources/schema/mapping_package_v3/models/mapping_package_v3_metadata.yaml")
-    
-    try:
-        generate_jsonld_context(
-            schema_yaml_path=schema_path,
-            output_directory=metadata_directory,
-            context_filename="context.jsonld"
-        )
-        logger.debug(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
-            package_source=mapping_package_folder_path,
-            message="Generated context.jsonld file"))
-    except Exception as e:
-        logger.warning(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
-            package_source=mapping_package_folder_path,
-            message=f"Failed to generate context.jsonld: {e}"))
-        # Don't fail the conversion if context generation fails
-        # The context file can be generated manually if needed
 
 
 def _convert_package_from_folder(from_version: str, to_version: str, mapping_package_folder_path: Path):
