+++ resources/schema/mapping_package_v3/models/context.jsonld
@@ -1,0 +1,87 @@
+{
+   "comments": {
+      "description": "Auto generated by LinkML jsonld context generator",
+      "generation_date": "2025-11-13T21:55:43",
+      "source": "mapping_package_v3_metadata.yaml"
+   },
+   "@context": {
+      "xsd": "http://www.w3.org/2001/XMLSchema#",
+      "linkml": "https://w3id.org/linkml/",
+      "mfy": "https://meaningfy.ws/mapping/",
+      "@vocab": "https://meaningfy.ws/mapping/",
+      "applicability_constraints": {
+         "@type": "mfy:ApplicabilityConstraints",
+         "@id": "applicability_constraints"
+      },
+      "created_at": {
+         "@type": "xsd:dateTime",
+         "@id": "created_at"
+      },
+      "description": {
+         "@id": "description"
+      },
+      "document_schema_version": {
+         "@id": "document_schema_version"
+      },
+      "document_schema_version_list": {
+         "@id": "document_schema_version_list"
+      },
+      "document_time_interval": {
+         "@type": "mfy:DateTimeInterval",
+         "@id": "document_time_interval"
+      },
+      "document_type_list": {
+         "@id": "document_type_list"
+      },
+      "document_version_range": {
+         "@type": "mfy:VersionRange",
+         "@id": "document_version_range"
+      },
+      "end": {
+         "@type": "xsd:dateTime",
+         "@id": "end"
+      },
+      "id": "@id",
+      "input_mime_types": {
+         "@id": "input_mime_types"
+      },
+      "mapping_suite_hash_digest": {
+         "@id": "mapping_suite_hash_digest"
+      },
+      "mapping_version": {
+         "@id": "mapping_version"
+      },
+      "max": {
+         "@id": "max"
+      },
+      "min": {
+         "@id": "min"
+      },
+      "model_version": {
+         "@id": "model_version"
+      },
+      "project_identifier": {
+         "@id": "project_identifier"
+      },
+      "start": {
+         "@type": "xsd:dateTime",
+         "@id": "start"
+      },
+      "title": {
+         "@id": "title"
+      },
+      "ApplicabilityConstraints": {
+         "@id": "ApplicabilityConstraints"
+      },
+      "DateTimeInterval": {
+         "@id": "DateTimeInterval"
+      },
+      "MappingPackageV3Metadata": {
+         "@id": "MappingPackageV3Metadata"
+      },
+      "VersionRange": {
+         "@id": "VersionRange"
+      }
+   }
+}
+
