+++ tests/unit/mapping_package_v3/services/test_convert_mapping_package_v2_to_v3.py
@@ -172,181 +172,3 @@ def test_convert_mapping_package_v2_to_v3_handles_both_dates(dummy_mapping_packa
         assert result.metadata.applicability_constraints.document_time_interval.end is not None
 
 
-def test_generate_jsonld_context_success(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context succeeds when gen-jsonld-context command works."""
-    import subprocess
-    from unittest.mock import Mock, patch
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    # Create a mock schema file
-    schema_path = tmp_path / "schema.yaml"
-    schema_path.write_text("id: TestSchema\n")
-    
-    output_dir = tmp_path / "output"
-    output_dir.mkdir()
-    
-    # Mock subprocess.run to return success
-    mock_result = Mock()
-    mock_result.stdout = '{"@context": {"test": "value"}}'
-    mock_result.stderr = ""
-    
-    with patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3.subprocess.run") as mock_run, \
-         patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3._find_project_root", return_value=tmp_path):
-        mock_run.return_value = mock_result
-        
-        result_path = generate_jsonld_context(
-            schema_yaml_path=schema_path,
-            output_directory=output_dir,
-            context_filename="context.jsonld"
-        )
-        
-        assert result_path == output_dir / "context.jsonld"
-        assert result_path.exists()
-        assert result_path.read_text() == '{"@context": {"test": "value"}}'
-        # Verify gen-jsonld-context was called directly (no poetry)
-        mock_run.assert_called_once()
-        call_args = mock_run.call_args[0][0]
-        assert call_args == ["gen-jsonld-context", str(schema_path)]
-
-
-def test_generate_jsonld_context_raises_file_not_found_error(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context raises FileNotFoundError when command is not in PATH."""
-    import subprocess
-    from unittest.mock import patch
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    # Create a mock schema file
-    schema_path = tmp_path / "schema.yaml"
-    schema_path.write_text("id: TestSchema\n")
-    
-    output_dir = tmp_path / "output"
-    output_dir.mkdir()
-    
-    with patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3.subprocess.run") as mock_run, \
-         patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3._find_project_root", return_value=tmp_path):
-        # Command not found in PATH
-        mock_run.side_effect = FileNotFoundError("gen-jsonld-context not found")
-        
-        with pytest.raises(FileNotFoundError):
-            generate_jsonld_context(
-                schema_yaml_path=schema_path,
-                output_directory=output_dir,
-                context_filename="context.jsonld"
-            )
-        
-        # Verify command was attempted once
-        mock_run.assert_called_once()
-
-
-def test_generate_jsonld_context_raises_error_when_schema_not_found(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context raises FileNotFoundError when schema file doesn't exist."""
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    schema_path = tmp_path / "nonexistent.yaml"
-    output_dir = tmp_path / "output"
-    output_dir.mkdir()
-    
-    with pytest.raises(FileNotFoundError) as excinfo:
-        generate_jsonld_context(
-            schema_yaml_path=schema_path,
-            output_directory=output_dir
-        )
-    
-    assert "Schema YAML file not found" in str(excinfo.value)
-
-
-def test_generate_jsonld_context_raises_called_process_error(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context raises CalledProcessError when command execution fails."""
-    import subprocess
-    from unittest.mock import patch
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    # Create a mock schema file
-    schema_path = tmp_path / "schema.yaml"
-    schema_path.write_text("id: TestSchema\n")
-    
-    output_dir = tmp_path / "output"
-    output_dir.mkdir()
-    
-    # Mock subprocess.run to fail
-    with patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3.subprocess.run") as mock_run, \
-         patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3._find_project_root", return_value=tmp_path):
-        # Command execution fails
-        mock_run.side_effect = subprocess.CalledProcessError(1, "gen-jsonld-context", stderr="command error")
-        
-        with pytest.raises(subprocess.CalledProcessError):
-            generate_jsonld_context(
-                schema_yaml_path=schema_path,
-                output_directory=output_dir
-            )
-        
-        # Verify command was attempted once
-        mock_run.assert_called_once()
-
-
-def test_generate_jsonld_context_creates_output_directory(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context creates output directory if it doesn't exist."""
-    import subprocess
-    from unittest.mock import Mock, patch
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    # Create a mock schema file
-    schema_path = tmp_path / "schema.yaml"
-    schema_path.write_text("id: TestSchema\n")
-    
-    # Output directory doesn't exist yet
-    output_dir = tmp_path / "output"
-    
-    # Mock subprocess.run to return success
-    mock_result = Mock()
-    mock_result.stdout = '{"@context": {}}'
-    mock_result.stderr = ""
-    
-    with patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3.subprocess.run") as mock_run, \
-         patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3._find_project_root", return_value=tmp_path):
-        mock_run.return_value = mock_result
-        
-        result_path = generate_jsonld_context(
-            schema_yaml_path=schema_path,
-            output_directory=output_dir
-        )
-        
-        # Verify directory was created
-        assert output_dir.exists()
-        assert output_dir.is_dir()
-        assert result_path.exists()
-
-
-def test_generate_jsonld_context_uses_custom_filename(tmp_path: Path) -> None:
-    """Test that generate_jsonld_context uses custom context filename when provided."""
-    import subprocess
-    from unittest.mock import Mock, patch
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3 import generate_jsonld_context
-    
-    # Create a mock schema file
-    schema_path = tmp_path / "schema.yaml"
-    schema_path.write_text("id: TestSchema\n")
-    
-    output_dir = tmp_path / "output"
-    output_dir.mkdir()
-    
-    # Mock subprocess.run to return success
-    mock_result = Mock()
-    mock_result.stdout = '{"@context": {}}'
-    mock_result.stderr = ""
-    
-    with patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3.subprocess.run") as mock_run, \
-         patch("mapping_suite_sdk.tools.services.convert_mapping_package_v2_to_v3._find_project_root", return_value=tmp_path):
-        mock_run.return_value = mock_result
-        
-        result_path = generate_jsonld_context(
-            schema_yaml_path=schema_path,
-            output_directory=output_dir,
-            context_filename="custom_context.jsonld"
-        )
-        
-        assert result_path == output_dir / "custom_context.jsonld"
-        assert result_path.exists()
-
-
-
