+++ tests/unit/tools/entrypoints/cli/test_convert_cli.py
@@ -554,94 +554,24 @@ def test_is_already_converted_v3_returns_false_for_invalid_metadata(
     assert result is False
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
-def test_serialise_mapping_package_v3_generates_context_jsonld(
-    mock_generate_context,
-    tmp_path: Path,
-    fixture_mapping_package_v3_model: MappingPackageV3
-) -> None:
-    """Test that _serialise_mapping_package generates context.jsonld for V3 packages."""
-    from mapping_suite_sdk.tools.entrypoints.cli.convert import Version
-    
-    # Mock the context generation to return a path
-    mock_generate_context.return_value = tmp_path / "metadata.jsonld" / "context.jsonld"
-    
-    # Serialise V3 package using enum value
-    _serialise_mapping_package(Version.V3, tmp_path, fixture_mapping_package_v3_model)
-    
-    # Verify context generation was called
-    mock_generate_context.assert_called_once()
-    call_kwargs = mock_generate_context.call_args[1]
-    assert call_kwargs["context_filename"] == "context.jsonld"
-    # Verify it was called with the metadata directory
-    metadata_dir = (tmp_path / fixture_mapping_package_v3_model.metadata.path).parent
-    assert call_kwargs["output_directory"] == metadata_dir
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
-def test_serialise_mapping_package_v3_lightweight_generates_context_jsonld(
-    mock_generate_context,
-    tmp_path: Path,
-    fixture_mapping_package_v3_model: MappingPackageV3
-) -> None:
-    """Test that _serialise_mapping_package generates context.jsonld for V3L packages."""
-    from mapping_suite_sdk.tools.entrypoints.cli.convert import Version
-    from mapping_suite_sdk.tools.services.convert_mapping_package_v3_to_v3_lightweight import convert_mapping_package_v3_to_v3_lightweight
-    
-    # Mock the context generation
-    mock_generate_context.return_value = tmp_path / "metadata.jsonld" / "context.jsonld"
-    
-    # Convert to lightweight and serialise using enum value
-    lightweight_package = convert_mapping_package_v3_to_v3_lightweight(fixture_mapping_package_v3_model)
-    _serialise_mapping_package(Version.V3L, tmp_path, lightweight_package)
-    
-    # Verify context generation was called
-    mock_generate_context.assert_called_once()
-    call_kwargs = mock_generate_context.call_args[1]
-    assert call_kwargs["context_filename"] == "context.jsonld"
-
-
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
-def test_serialise_mapping_package_v3_continues_on_context_generation_failure(
-    mock_generate_context,
-    tmp_path: Path,
-    fixture_mapping_package_v3_model: MappingPackageV3,
-    caplog
-) -> None:
-    """Test that _serialise_mapping_package continues even if context generation fails."""
-    from mapping_suite_sdk.tools.entrypoints.cli.convert import Version
-    
-    # Mock context generation to raise an error
-    mock_generate_context.side_effect = RuntimeError("Failed to generate context")
-    
-    # Serialise should still succeed (context generation failure is logged but doesn't fail)
-    _serialise_mapping_package(Version.V3, tmp_path, fixture_mapping_package_v3_model)
-    
-    # Verify metadata file was still created
-    assert (tmp_path / fixture_mapping_package_v3_model.metadata.path).exists()
-    # Verify warning was logged
-    assert "Failed to generate context.jsonld" in caplog.text
-
-
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert.is_mapping_package_already_converted", return_value=False)
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._load_mapping_package_from_folder")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._convert_mapping_package")
-def test_convert_from_package_generates_context_jsonld(
+def test_convert_from_package_serialises_package(
     mock_convert,
     mock_load,
     mock_is_already_converted,
-    mock_generate_context,
     typer_cli_runner: CliRunner,
     tmp_path: Path,
     dummy_mapping_package_v2_model,
     fixture_mapping_package_v3_model: MappingPackageV3
 ) -> None:
-    """Test that from-package command generates context.jsonld during conversion."""
+    """Test that from-package command serialises the converted package."""
     # Setup mocks to allow real serialization to run
     mock_load.return_value = dummy_mapping_package_v2_model
     mock_convert.return_value = fixture_mapping_package_v3_model
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
     
     package_path = tmp_path / "test_package"
     package_path.mkdir()
@@ -652,26 +582,23 @@ def test_convert_from_package_generates_context_jsonld(
     )
     
     assert result.exit_code == 0
-    # Verify context generation was called during serialisation
-    assert mock_generate_context.called
+    # Verify metadata file was created
+    assert (package_path / fixture_mapping_package_v3_model.metadata.path).exists()
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert.is_mapping_package_already_converted", return_value=False)
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._load_mapping_package_from_folder")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._convert_mapping_package")
-def test_convert_from_folder_generates_context_jsonld_for_each_package(
+def test_convert_from_folder_serialises_each_package(
     mock_convert,
     mock_load,
     mock_is_already_converted,
-    mock_generate_context,
     typer_cli_runner: CliRunner,
     tmp_path: Path,
     dummy_mapping_package_v2_model,
-    fixture_mapping_package_v3_model: MappingPackageV3,
-    caplog
+    fixture_mapping_package_v3_model: MappingPackageV3
 ) -> None:
-    """Test that from-folder command generates context.jsonld for each converted package."""
+    """Test that from-folder command serialises each converted package."""
     folder_path = tmp_path / "packages"
     folder_path.mkdir()
     
@@ -684,7 +611,6 @@ def test_convert_from_folder_generates_context_jsonld_for_each_package(
     # Setup mocks to allow real serialization to run
     mock_load.return_value = dummy_mapping_package_v2_model
     mock_convert.return_value = fixture_mapping_package_v3_model
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
     
     result = typer_cli_runner.invoke(
         mssdk_cli_convert_subcommand,
@@ -695,13 +621,12 @@ def test_convert_from_folder_generates_context_jsonld_for_each_package(
     # Verify conversion was called for each package
     assert mock_load.call_count == 2
     assert mock_convert.call_count == 2
-    # Context generation is called within the conversion process for each package
-    assert mock_generate_context.call_count == 2
+    # Verify metadata files were created for each package
+    assert (package1 / fixture_mapping_package_v3_model.metadata.path).exists()
+    assert (package2 / fixture_mapping_package_v3_model.metadata.path).exists()
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 def test_serialise_mapping_package_v3_removes_old_metadata_json(
-    mock_generate_context,
     tmp_path: Path,
     fixture_mapping_package_v3_model: MappingPackageV3
 ) -> None:
@@ -713,9 +638,6 @@ def test_serialise_mapping_package_v3_removes_old_metadata_json(
     old_metadata_path.write_text('{"identifier": "old_v2_package"}')
     assert old_metadata_path.exists(), "Old metadata.json should exist before conversion"
     
-    # Mock context generation
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
-    
     # Serialise V3 package
     _serialise_mapping_package(Version.V3, tmp_path, fixture_mapping_package_v3_model)
     
@@ -725,9 +647,7 @@ def test_serialise_mapping_package_v3_removes_old_metadata_json(
     assert (tmp_path / fixture_mapping_package_v3_model.metadata.path).exists(), "New metadata.jsonld should exist"
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 def test_serialise_mapping_package_v3_lightweight_removes_old_metadata_json(
-    mock_generate_context,
     tmp_path: Path,
     fixture_mapping_package_v3_model: MappingPackageV3
 ) -> None:
@@ -740,9 +660,6 @@ def test_serialise_mapping_package_v3_lightweight_removes_old_metadata_json(
     old_metadata_path.write_text('{"identifier": "old_v2_package"}')
     assert old_metadata_path.exists(), "Old metadata.json should exist before conversion"
     
-    # Mock context generation
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
-    
     # Convert to lightweight and serialise
     lightweight_package = convert_mapping_package_v3_to_v3_lightweight(fixture_mapping_package_v3_model)
     _serialise_mapping_package(Version.V3L, tmp_path, lightweight_package)
@@ -753,9 +670,7 @@ def test_serialise_mapping_package_v3_lightweight_removes_old_metadata_json(
     assert (tmp_path / lightweight_package.metadata.path).exists(), "New metadata.jsonld should exist"
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 def test_serialise_mapping_package_v3_lightweight_removes_conceptual_mapping_file(
-    mock_generate_context,
     tmp_path: Path,
     fixture_mapping_package_v3_model: MappingPackageV3
 ) -> None:
@@ -770,9 +685,6 @@ def test_serialise_mapping_package_v3_lightweight_removes_conceptual_mapping_fil
     conceptual_mapping_path.write_bytes(b"fake xlsx content")
     assert conceptual_mapping_path.exists(), "conceptual_mappings.xlsx should exist before conversion"
     
-    # Mock context generation
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
-    
     # Convert to lightweight and serialise
     lightweight_package = convert_mapping_package_v3_to_v3_lightweight(fixture_mapping_package_v3_model)
     _serialise_mapping_package(Version.V3L, tmp_path, lightweight_package)
@@ -783,7 +695,6 @@ def test_serialise_mapping_package_v3_lightweight_removes_conceptual_mapping_fil
     assert (tmp_path / lightweight_package.metadata.path).exists(), "New metadata.jsonld should exist"
 
 
-@patch("mapping_suite_sdk.tools.entrypoints.cli.convert.generate_jsonld_context")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert.is_mapping_package_already_converted", return_value=False)
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._load_mapping_package_from_folder")
 @patch("mapping_suite_sdk.tools.entrypoints.cli.convert._convert_mapping_package")
@@ -791,7 +702,6 @@ def test_convert_from_package_removes_old_metadata_json(
     mock_convert,
     mock_load,
     mock_is_already_converted,
-    mock_generate_context,
     typer_cli_runner: CliRunner,
     tmp_path: Path,
     dummy_mapping_package_v2_model,
@@ -809,7 +719,6 @@ def test_convert_from_package_removes_old_metadata_json(
     # Setup mocks to allow real serialization to run
     mock_load.return_value = dummy_mapping_package_v2_model
     mock_convert.return_value = fixture_mapping_package_v3_model
-    mock_generate_context.return_value = tmp_path / "context.jsonld"
     
     result = typer_cli_runner.invoke(
         mssdk_cli_convert_subcommand,
